# api-doc.json
{
  "swagger": "2.0",
  "info": {
    "title": "Survey Node API",
    "version": "1.0.0",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/survey_questions/": {
      "get": {
        "tags": [
          "Questions"
        ],
        "summary": "Get All Questions",
        "description": "",
        "operationId": "getQuestions",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Questions"
          }
        }
      },
      "post": {
        "tags": [
          "Questions"
        ],
        "summary": "Delete Questions with IDs",
        "description": "",
        "operationId": "deleteMultipleQuestions",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "ids": {
                  "type": "array",
                  "items": {
                    "type": "number"
                  }
                }
              },
              "required": [
                "ids"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Questions"
          }
        }
      }
    },
    "/api/survey_questions/create": {
      "post": {
        "tags": [
          "Questions"
        ],
        "summary": "Create Question",
        "description": "",
        "operationId": "createQuestion",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "question": {
                  "type": "string"
                },
                "survey_id": {
                  "type": "number"
                }
              },
              "required": [
                "question",
                "survey_id"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Questions"
          }
        }
      }
    },
    "/api/survey_questions/{id}": {
      "get": {
        "tags": [
          "Questions"
        ],
        "summary": "Get Question",
        "description": "",
        "operationId": "getQuestion",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "number"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Questions"
          }
        }
      },
      "put": {
        "tags": [
          "Questions"
        ],
        "summary": "Update Question",
        "description": "",
        "operationId": "updateQuestion",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "number"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "question": {
                  "type": "string"
                }
              },
              "required": [
                "question"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Questions"
          }
        }
      },
      "delete": {
        "tags": [
          "Questions"
        ],
        "summary": "Delete Question",
        "description": "",
        "operationId": "deleteQuestion",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "number"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Questions"
          }
        }
      }
    },
...
    "/api/surveys/": {
      "get": {
        "tags": [
          "Surveys"
        ],
        "summary": "Get All Surveys",
        "description": "",
        "operationId": "getSurveys",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Surveys"
          }
        }
      },
      "post": {
        "tags": [
          "Surveys"
        ],
        "summary": "Multipurpose endpoint to create or update a survey, its corresponding questions, and options in one operation",
        "description": "",
        "operationId": "createOrUpdateSurveys",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "survey_name": {
                  "type": "string"
                },
                "survey_description": {
                  "type": "string"
                },
                "is_active": {
                  "type": "boolean"
                },
                "questions": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer"
                      },
                      "question": {
                        "type": "string"
                      },
                      "input_type_id": {
                        "type": "integer"
                      },
                      "admin_id": {
                        "type": "integer"
                      },
                      "options": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "integer"
                            },
                            "label": {
                              "type": "string"
                            },
                            "question_id": {
                              "type": "integer"
                            },
                            "admin_id": {
                              "type": "integer"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              },
              "required": [
                "survey_name",
                "survey_description",
                "is_active"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Surveys"
          }
        }
      }
    },
    "/api/surveys/create": {
      "post": {
        "tags": [
          "Surveys"
        ],
        "summary": "Multipurpose endpoint to create or update a survey, its corresponding questions, and options in one operation",
        "description": "",
        "operationId": "createOrUpdateSurveys",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "survey_name": {
                  "type": "string"
                },
                "survey_description": {
                  "type": "string"
                },
                "is_active": {
                  "type": "boolean"
                },
                "questions": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer"
                      },
                      "question": {
                        "type": "string"
                      },
                      "input_type_id": {
                        "type": "integer"
                      },
                      "admin_id": {
                        "type": "integer"
                      },
                      "options": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "integer"
                            },
                            "label": {
                              "type": "string"
                            },
                            "question_id": {
                              "type": "integer"
                            },
                            "admin_id": {
                              "type": "integer"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              },
              "required": [
                "survey_name",
                "survey_description"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Surveys"
          }
        }
      }
    },
...
    "/api/surveys/{id}/questions": {
      "get": {
        "tags": [
          "Surveys"
        ],
        "summary": "Get Survey Questions",
        "description": "",
        "operationId": "getSurveyQuestions",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Surveys"
          }
        }
      }
    },
...
    "/api/surveys/{id}/{id}": {
      "get": {
        "tags": [
          "Surveys"
        ],
        "summary": "Get Survey",
        "description": "",
        "operationId": "getSurvey",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "Surveys"
          }
        }
      },
      "put": {
        "tags": [
          "Surveys"
        ],
        "summary": "Update specific survey",
        "description": "",
        "operationId": "updateSurveys",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "number"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "survey_name": {
                  "type": "string"
                },
                "survey_description": {
                  "type": "string"
                },
                "is_active": {
                  "type": "boolean"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Surveys"
          }
        }
      },
      "delete": {
        "tags": [
          "Surveys"
        ],
        "summary": "Delete Survey",
        "description": "",
        "operationId": "deleteSurvey",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "number"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Survey deleted"
          }
        }
      }
    }
  }
}